+++ app/src/main/java/com/samkit/costcircle/ui/screens/groups/states/GroupsEvent.kt
@@ -3,5 +3,5 @@ package com.samkit.costcircle.ui.screens.groups.states
 sealed interface GroupsEvent {
     data object LoadGroups : GroupsEvent
     data object Retry : GroupsEvent
-    data class GroupClicked(val groupId: Int, val name: String) : GroupsEvent
+    data class GroupClicked(val groupId: Long, val name: String) : GroupsEvent
 }
